+++ Changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added [BLEURT](models/sellam2020/Readme.md)
 - Added [BERTScore](models/zhang2020/Readme.md)
+- Added [BLEU and SentBLEU](models/papineni2002/Readme.md)
 
 ## [v0.0.2](https://github.com/danieldeutsch/repro/repro/tag/0.0.2) - 2021-07-30
 ### Added