+++ Changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [v0.0.3](https://github.com/danieldeutsch/repro/releases/tag/v0.0.3) - 2021-08-04
 ### Added
 - Added [BLEURT](models/sellam2020/Readme.md)
 - Added [BERTScore](models/zhang2020/Readme.md)
@@ -17,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed the QAEval interface to match other text generation metrics.
 The backend was also changed to not rely on SacreROUGE.
 
-## [v0.0.2](https://github.com/danieldeutsch/repro/repro/tag/0.0.2) - 2021-07-30
+## [v0.0.2](https://github.com/danieldeutsch/repro/releases/tag/v0.0.2) - 2021-07-30
 ### Added
 - Added a `RecipeGenerationModel` class
 - Added a [recipe generation model](models/dugan2020/Readme.md) from [Dugan et al. (2020)](https://arxiv.org/abs/2010.03070)
@@ -40,6 +42,6 @@ See [here](models/lin2004/Readme.md)
 - Renamed the `--model-args`, `--dataset-reader-args`, and `--output-write-args` `predict` arguments to `--model-kwargs`, `--dataset-reader-kwargs`, and `--output-write-kwargs`.
 - Renamed the `--output-file` argument in `predict` to `--output` to allow for output files or directories.
 
-## [v0.0.1](https://github.com/danieldeutsch/repro/repro/tag/0.0.1) - 2021-07-22
+## [v0.0.1](https://github.com/danieldeutsch/repro/releases/tag/v0.0.1) - 2021-07-22
 ### Added
 - Initial prototype of the library with `setup` and `predict` commands as well as implementations of [Gupta et al. (2020)](models/gupta2020/Readme.md), [Lewis et al. (2020)](models/lewis2020/Readme.md), and [Liu & Lapata (2019)](models/liu2019/Readme.md).